Rename Package & Minor Changes
