+++ src/main/java/com/github/githubbers/mainPlug.java
@@ -1,11 +1,10 @@
-package com.github.macasyraf;
+package com.github.githubbers;
 
 public class mainPlug {
 
     public static void main(String[] args) {
 
         new mainPlug().getPlug();
-        repoPlug.getLink(); // and also this one.
     }
 
     private void getPlug() {
